+++ spec/markovian/chain_spec.rb
@@ -3,10 +3,10 @@ require 'spec_helper'
 module Markovian
   RSpec.describe Chain do
     let(:chain) { Chain.new }
-    let(:word) { Tokeneyes::Word.new(Faker::Lorem.word) }
-    let(:next_word) { Tokeneyes::Word.new(Faker::Lorem.word) }
-    let(:previous_word) { Tokeneyes::Word.new(Faker::Lorem.word) }
-    let(:phrase_association) { Tokeneyes::Word.new(Faker::Lorem.word) }
+    let(:word) { Faker::Lorem.word }
+    let(:next_word) { Faker::Lorem.word }
+    let(:previous_word) { Faker::Lorem.word }
+    let(:phrase_association) { Faker::Lorem.word }
 
     describe "#word_entry" do
       it "returns the dictionary entry for the word" do
@@ -14,70 +14,85 @@ module Markovian
       end
     end
 
-    describe "#next_word" do
-      it "returns no values when empty" do
-        expect(chain.next_word(word)).to be_nil
+    describe "#lengthen / #next_word" do
+      shared_examples_for :lengthening_and_getting_next_word do
+        it "returns no values when empty" do
+          expect(chain.next_word(word)).to be_nil
+        end
+
+        context "when populated" do
+          context "with a single-word match" do
+            before :each do
+              chain.lengthen(word, next_word: next_word)
+            end
+
+            it "returns the next word when looking up by word" do
+              expect(chain.next_word(word)).to eq(next_word.to_s)
+            end
+
+            it "returns the next word when looking up by phrase" do
+              expect(chain.next_word(word, previous_word: previous_word)).to eq(next_word.to_s)
+            end
+
+            it "populates the the entry's data" do
+              chain.next_word(word, previous_word: previous_word)
+              expect(chain.word_entry(word).occurrences).to eq(1)
+            end
+          end
+
+          context "with a phrase match and a single word match" do
+            before :each do
+              chain.lengthen(word, next_word: next_word)
+              chain.lengthen(word, previous_word: previous_word, next_word: phrase_association)
+            end
+
+            it "samples the results when given a single word", temporary_srand: 12 do
+              # jruby's srand returns something slightly different from MRI's, so we need to
+              # hard-code two results
+              if RUBY_PLATFORM == "java"
+                results = [
+                  next_word, phrase_association, phrase_association, next_word, phrase_association, next_word, next_word
+                ]
+              else
+                results = [
+                  next_word, phrase_association, phrase_association, next_word, next_word, next_word, phrase_association
+                ]
+              end
+              expect(7.times.map { chain.next_word(word) }).to eq(results.map(&:to_s))
+            end
+
+            it "returns only the next match when looking up by phrase" do
+              expect(7.times.map { chain.next_word(word, previous_word: previous_word) }).to eq([phrase_association.to_s] * 7)
+            end
+          end
+
+          context "with only a phrase match" do
+            before :each do
+              chain.lengthen(word, previous_word: previous_word, next_word: phrase_association)
+            end
+
+            it "returns only the next phrase match when looking up by word" do
+              expect(7.times.map { chain.next_word(word) }).to eq([phrase_association.to_s] * 7)
+            end
+
+            it "returns only the next match when looking up by phrase" do
+              expect(7.times.map { chain.next_word(word, previous_word: previous_word) }).to eq([phrase_association.to_s] * 7)
+            end
+          end
+        end
       end
 
-      context "when populated" do
-        context "with a single-word match" do
-          before :each do
-            chain.lengthen(word, next_word: next_word)
-          end
+      context "when given strings" do
+        it_should_behave_like :lengthening_and_getting_next_word
+      end
 
-          it "returns the next word when looking up by word" do
-            expect(chain.next_word(word)).to eq(next_word.to_s)
-          end
+      context "when given Tokeneyes::Word entries" do
+        let(:word) { Tokeneyes::Word.new(Faker::Lorem.word) }
+        let(:next_word) { Tokeneyes::Word.new(Faker::Lorem.word) }
+        let(:previous_word) { Tokeneyes::Word.new(Faker::Lorem.word) }
+        let(:phrase_association) { Tokeneyes::Word.new(Faker::Lorem.word) }
 
-          it "returns the next word when looking up by phrase" do
-            expect(chain.next_word(word, previous_word: previous_word)).to eq(next_word.to_s)
-          end
-
-          it "populates the the entry's data" do
-            chain.next_word(word, previous_word: previous_word)
-            expect(chain.word_entry(word).occurrences).to eq(1)
-          end
-        end
-
-        context "with a phrase match and a single word match" do
-          before :each do
-            chain.lengthen(word, next_word: next_word)
-            chain.lengthen(word, previous_word: previous_word, next_word: phrase_association)
-          end
-
-          it "samples the results when given a single word", temporary_srand: 12 do
-            # jruby's srand returns something slightly different from MRI's, so we need to
-            # hard-code two results
-            if RUBY_PLATFORM == "java"
-              results = [
-                next_word, phrase_association, phrase_association, next_word, phrase_association, next_word, next_word
-              ]
-            else
-              results = [
-                next_word, phrase_association, phrase_association, next_word, next_word, next_word, phrase_association
-              ]
-            end
-            expect(7.times.map { chain.next_word(word) }).to eq(results.map(&:to_s))
-          end
-
-          it "returns only the next match when looking up by phrase" do
-            expect(7.times.map { chain.next_word(word, previous_word: previous_word) }).to eq([phrase_association.to_s] * 7)
-          end
-        end
-
-        context "with only a phrase match" do
-          before :each do
-            chain.lengthen(word, previous_word: previous_word, next_word: phrase_association)
-          end
-
-          it "returns only the next phrase match when looking up by word" do
-            expect(7.times.map { chain.next_word(word) }).to eq([phrase_association.to_s] * 7)
-          end
-
-          it "returns only the next match when looking up by phrase" do
-            expect(7.times.map { chain.next_word(word, previous_word: previous_word) }).to eq([phrase_association.to_s] * 7)
-          end
-        end
+        it_should_behave_like :lengthening_and_getting_next_word
       end
     end