+++ changelog.md
@@ -2,6 +2,9 @@
 
 ## 0.3.0
 
+* TextBuilder#construct now includes seed text by default (instead of via opt-in)
+* Add Chain#word_entry to allow access to word data
+* Properly collect metadata about words (previously collected next_word's data)
 * Refactor Dictionary to provide access to entries, removing a lot of method duplication
 
 ## 0.2.9